back-end/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-dotenv
requests
pydantic
typing-extensions
aiohttp
pycryptodome
llama-index-vector-stores-postgres
beautifulsoup4
boto3
python-docx
docx
botocore
langdetect
llama-index
llama-cloud-services
llama-parse
PyMuPDF
pyparsing
python-dotenv
sqlmodel
psycopg2-binary
websockets
Pillow
